# Ubli-Front/src/api/schemas/pontoAcessibilidade.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Ubli - Acessibilidade urbana
 */

export interface PontoAcessibilidade {
  descricao?: string;
  classificacaoLocal?: number;
  latitude?: number;
  longitude?: number;
  idUsuario?: string;
  fotosLocal?: string[];
}