# Ubli-Front/src/api/schemas/pontoAcessibilidadeDTO.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Ubli - Acessibilidade urbana
 */

export interface PontoAcessibilidadeDTO {
  descricao?: string;
  classificacaoLocal?: number;
  latitude?: number;
  longitude?: number;
  idUsuario?: string;
}