# Ubli-Front/src/api/schemas/usuario.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Ubli - Acessibilidade urbana
 */

export interface Usuario {
  nome?: string;
  email?: string;
  senha?: string;
  cpf?: string;
  fotoPerfil?: string[];
}